requirements.txt
streamlit
sqlalchemy
pydantic
pytest
langgraph
pdf2image
pytesseract
pypdf
google-genai
python-dotenv
